+++ src/components/loader.js
@@ -1,0 +1,7 @@
+function Loader() {
+    return (
+      <p className="message-display">Loading...</p>
+    )
+  }
+
+  export default Loader;
