+++ src/components/prayerBox.js
@@ -1,4 +1,3 @@
-import React from "react";
 import { useState } from "react";
 import supabase from "../supabase";
